refactor(G1): split resolver/mod.rs into fd, import_edit, hierarchy sub-modules
- resolver/fd.rs: package_prefix, import_file_stems, fd_find_and_parse,
  fd_search_by_full_path_pattern, fd_search_file, parse_fd_hits
- resolver/import_edit.rs: already_imported, import_insertion_line, make_import_edit
- resolver/hierarchy.rs: walk_hierarchy, HierarchyWalker, supertype_targets,
  super_names_for_class

mod.rs keeps the resolution chain (~520 lines, down from 1095).
External consumers unchanged — re-exports preserve public API paths.
830 tests pass, 0 new clippy warnings.
