+++ src/resolver/mod.rs
@@ -23,17 +23,21 @@ use std::path::Path;
 use std::process::Command;
 use std::sync::Arc;
 
-use tower_lsp::lsp_types::{Location, Range, TextEdit, Url};
+use tower_lsp::lsp_types::{Location, Url};
 
 use crate::indexer::Indexer;
 use crate::parser::parse_by_extension;
 use crate::rg::{build_rg_pattern, parse_rg_line, rg_find_definition};
-use crate::types::{CallerContext, FileData, ImportEntry};
-use crate::LinesExt;
+use crate::types::{CallerContext, FileData};
 use crate::StrExt;
 
+use fd::{fd_find_and_parse, package_prefix};
+
 pub(crate) mod complete;
+mod fd;
 pub(crate) mod find;
+mod hierarchy;
+mod import_edit;
 pub(crate) mod infer;
 #[cfg(test)]
 mod tests;
@@ -44,6 +48,8 @@ pub(crate) use complete::is_annotation_context;
 pub(crate) use complete::{
     complete_symbol, complete_symbol_with_context, symbols_from_uri_as_completions_pub,
 };
+pub(crate) use hierarchy::walk_hierarchy;
+pub(crate) use import_edit::{already_imported, import_insertion_line, make_import_edit};
 pub(crate) use infer::{
     extract_collection_element_type, infer_receiver_type, infer_variable_type_raw, ReceiverKind,
     ReceiverType,
@@ -54,6 +60,8 @@ pub(crate) use infer::{
 pub(crate) use complete::{
     complete_bare, complete_dot, is_screaming_snake, match_score, COMPLETION_CAP,
 };
+#[cfg(test)]
+use fd::import_file_stems;
 #[cfg(test)]
 pub(crate) use infer::{
     find_declaration_range_in_lines, infer_type_in_lines, infer_type_in_lines_raw,
@@ -76,114 +84,6 @@ pub(crate) fn ensure_file_data(idx: &Indexer, uri: &Url) -> Option<Arc<FileData>
     Some(Arc::new(parse_by_extension(uri.path(), &content)))
 }
 
-/// Walk the class hierarchy starting from `start_class`, collecting items at each level.
-/// `T` is what the visitor produces per symbol. `max_depth` prevents infinite loops.
-pub(crate) fn walk_hierarchy<'a, T, F>(
-    idx: &'a Indexer,
-    start_class: &str,
-    start_uri: &str,
-    caller: CallerContext<'a>,
-    max_depth: usize,
-    collect: F,
-) -> Vec<T>
-where
-    F: Fn(&Indexer, &str, &str, CallerContext<'_>) -> Vec<T>,
-{
-    let mut walker = HierarchyWalker {
-        idx,
-        caller,
-        max_depth,
-        collect,
-        visited: HashSet::from([(start_uri.to_owned(), start_class.to_owned())]),
-        items: Vec::new(),
-    };
-    walker.recurse(start_class, start_uri, 0);
-    walker.items
-}
-
-struct HierarchyWalker<'a, T, F>
-where
-    F: Fn(&Indexer, &str, &str, CallerContext<'_>) -> Vec<T>,
-{
-    idx: &'a Indexer,
-    caller: CallerContext<'a>,
-    max_depth: usize,
-    collect: F,
-    visited: HashSet<(String, String)>,
-    items: Vec<T>,
-}
-
-impl<'a, T, F> HierarchyWalker<'a, T, F>
-where
-    F: Fn(&Indexer, &str, &str, CallerContext<'_>) -> Vec<T>,
-{
-    fn recurse(&mut self, class_name: &str, class_uri: &str, depth: usize) {
-        if depth >= self.max_depth {
-            return;
-        }
-
-        for (super_name, super_uri) in supertype_targets(self.idx, class_name, class_uri) {
-            if !self.visited.insert((super_uri.clone(), super_name.clone())) {
-                continue;
-            }
-            self.items.extend((self.collect)(
-                self.idx,
-                &super_name,
-                &super_uri,
-                self.caller,
-            ));
-            self.recurse(&super_name, &super_uri, depth + 1);
-        }
-    }
-}
-
-fn supertype_targets(idx: &Indexer, class_name: &str, class_uri: &str) -> Vec<(String, String)> {
-    let Ok(uri) = Url::parse(class_uri) else {
-        return vec![];
-    };
-    let Some(file_data) = ensure_file_data(idx, &uri) else {
-        return vec![];
-    };
-
-    super_names_for_class(&file_data, class_name)
-        .into_iter()
-        .flat_map(|super_name| {
-            resolve_symbol_inner(idx, &super_name, &uri, false)
-                .into_iter()
-                .map(move |loc| (super_name.clone(), loc.uri.to_string()))
-        })
-        .collect()
-}
-
-fn super_names_for_class(file_data: &FileData, class_name: &str) -> Vec<String> {
-    if class_name.is_empty() {
-        return file_data
-            .supers
-            .iter()
-            .map(|(_, name, _)| name.clone())
-            .collect();
-    }
-
-    let class_line = file_data
-        .symbols
-        .iter()
-        .find(|symbol| symbol.name == class_name)
-        .map(|symbol| symbol.selection_start());
-    match class_line {
-        Some(line) => file_data
-            .supers
-            .iter()
-            .filter(|(super_line, _, _)| *super_line == line)
-            .map(|(_, name, _)| name.clone())
-            .collect(),
-        None => file_data
-            .supers
-            .iter()
-            .map(|(_, name, _)| name.clone())
-            .collect(),
-    }
-}
-
 // ─── auto-import helpers ──────────────────────────────────────────────────────
 
 /// Return all importable FQNs for a simple symbol name (e.g. "Composable").
@@ -192,83 +92,6 @@ pub(crate) fn fqns_for_name(idx: &Indexer, name: &str) -> Vec<String> {
         .read()
         .map(|m| m.get(name).cloned().unwrap_or_default())
         .unwrap_or_default()
-}
-
-/// True if `fqn` is already usable in the file without an additional import:
-/// - exact non-alias import: `import pkg.Name` where local_name == last segment
-/// - star import covering the package: `import pkg.*`
-pub(crate) fn already_imported(fqn: &str, imports: &[ImportEntry]) -> bool {
-    let last_seg = fqn.rsplit('.').next().unwrap_or(fqn);
-    let pkg = match fqn.rfind('.') {
-        Some(i) => &fqn[..i],
-        None => "",
-    };
-    imports.iter().any(|imp| {
-        if imp.is_star {
-            imp.full_path == pkg
-        } else {
-            // Only count as "already imported" when not aliased to a different name.
-            imp.full_path == fqn && imp.local_name == last_seg
-        }
-    })
-}
-
-/// Find the line number after which to insert a new import statement.
-/// Returns the 0-based line index of the *new* import line (the line we'll insert before).
-/// Priority: after the last existing `import` line; else after the `package` line; else line 0.
-pub(crate) fn import_insertion_line(lines: &[String]) -> u32 {
-    // Find the last import line.
-    let last_import = lines
-        .iter()
-        .enumerate()
-        .rev()
-        .find(|(_, l)| l.trim_start().starts_with("import "))
-        .map(|(i, _)| i);
-    if let Some(i) = last_import {
-        return (i + 1) as u32;
-    }
-    // No imports — insert after package declaration (with a blank line gap).
-    let pkg_line = lines
-        .iter()
-        .enumerate()
-        .find(|(_, l)| l.trim_start().starts_with("package "))
-        .map(|(i, _)| i);
-    if let Some(i) = pkg_line {
-        return (i + 1) as u32;
-    }
-    0
-}
-
-/// Build a TextEdit that inserts `import {fqn}\n` at the correct position.
-pub(crate) fn make_import_edit(fqn: &str, lines: &[String], needs_semicolon: bool) -> TextEdit {
-    let line = lines.import_insertion_line();
-    // When inserting right after the package line (no existing imports), add a blank line.
-    let needs_blank = line > 0
-        && lines
-            .get((line - 1) as usize)
-            .map(|l| l.trim_start().starts_with("package "))
-            .unwrap_or(false)
-        && lines
-            .get(line as usize)
-            .map(|l| !l.trim().is_empty())
-            .unwrap_or(false);
-    let stmt = if needs_semicolon {
-        format!("import {fqn};")
-    } else {
-        format!("import {fqn}")
-    };
-    let new_text = if needs_blank {
-        format!("\n{stmt}\n")
-    } else {
-        format!("{stmt}\n")
-    };
-    TextEdit {
-        range: Range {
-            start: tower_lsp::lsp_types::Position { line, character: 0 },
-            end: tower_lsp::lsp_types::Position { line, character: 0 },
-        },
-        new_text,
-    }
 }
 
 /// Resolve `name` as seen from `from_uri`, returning all known definition
@@ -656,191 +479,6 @@ fn resolve_via_imports(idx: &Indexer, name: &str, uri: &Url) -> Vec<Location> {
         }
     }
     vec![]
-}
-
-/// Derive the Kotlin package from an import path by taking all dot-separated
-/// segments that start with a lowercase letter (package convention).
-///
-/// `cz.moneta.app.AccountPickerContract.Event` → `"cz.moneta.app"`
-fn package_prefix(import_path: &str) -> String {
-    import_path
-        .split('.')
-        .take_while(|s| s.starts_with_lowercase())
-        .collect::<Vec<_>>()
-        .join(".")
-}
-
-/// Uppercase segment stems in priority order — outer class first.
-///
-/// `com.example.OuterClass.InnerClass` → `["OuterClass", "InnerClass"]`
-/// `com.example.Foo`                   → `["Foo"]`
-fn import_file_stems(import_path: &str) -> Vec<String> {
-    let upper: Vec<&str> = import_path
-        .split('.')
-        .filter(|s| s.starts_with_uppercase())
-        .collect();
-    match upper.as_slice() {
-        [] => vec![],
-        [only] => vec![only.to_string()],
-        [.., par, lst] => vec![par.to_string(), lst.to_string()],
-    }
-}
-
-/// Find and synchronously parse the file most likely to contain `symbol_name`.
-///
-/// Search strategy (fastest-first):
-///   1. fd `--full-path` regex derived from the import's package dir + filename —
-///      extremely precise; handles multi-module projects where files live in
-///      subdirs like `app/src/main/java/cz/moneta/…/EProductScreen.java`
-///   2. Fallback: global fd by filename only (handles non-standard layouts)
-fn fd_find_and_parse(
-    symbol_name: &str,
-    full_import_path: &str,
-    root: Option<&Path>,
-    matcher: Option<&crate::rg::IgnoreMatcher>,
-) -> Vec<Location> {
-    let pkg = package_prefix(full_import_path);
-    let expected_pkg = if pkg.is_empty() {
-        None
-    } else {
-        Some(pkg.as_str())
-    };
-    let pkg_dir = pkg.replace('.', "/");
-
-    let ext_alt = crate::rg::SOURCE_EXTENSIONS.join("|");
-    for stem in import_file_stems(full_import_path) {
-        // Strategy 1: precise full-path regex including the package directory.
-        // e.g. ".*/cz/moneta/data/compat/enums/product/EProductScreen\.(kt|java|swift)$"
-        if let Some(root) = root {
-            let pat = if pkg_dir.is_empty() {
-                format!(r"{stem}\.({ext_alt})$")
-            } else {
-                format!(r".*/{pkg_dir}/{stem}\.({ext_alt})$")
-            };
-            let locs = fd_search_by_full_path_pattern(&pat, symbol_name, expected_pkg, root);
-            let locs = match matcher {
-                Some(m) => m.filter_locs(locs),
-                None => locs,
-            };
-            if !locs.is_empty() {
-                return locs;
-            }
-        }
-
-        // Strategy 2: global filename-only search (fallback for flat / non-standard layouts).
-        for ext in crate::rg::SOURCE_EXTENSIONS {
-            let locs = fd_search_file(&format!("{stem}.{ext}"), symbol_name, expected_pkg, root);
-            let locs = match matcher {
-                Some(m) => m.filter_locs(locs),
-                None => locs,
-            };
-            if !locs.is_empty() {
-                return locs;
-            }
-        }
-    }
-    vec![]
-}
-
-/// fd `--full-path <regex>` — searches `root` for files whose absolute path
-/// matches `pattern`.  Parses each hit and returns locations for `symbol_name`.
-fn fd_search_by_full_path_pattern(
-    pattern: &str,
-    symbol_name: &str,
-    expected_pkg: Option<&str>,
-    root: &Path,
-) -> Vec<Location> {
-    let Some(root_str) = root.to_str() else {
-        return vec![];
-    };
-    let out = match std::process::Command::new("fd")
-        .args([
-            "--type",
-            "f",
-            "--absolute-path",
-            "--full-path",
-            pattern,
-            root_str,
-        ])
-        .output()
-    {
-        Ok(o) if o.status.success() => o,
-        _ => return vec![],
-    };
-    parse_fd_hits(&out.stdout, symbol_name, expected_pkg)
-}
-
-fn fd_search_file(
-    file_name: &str,
-    symbol_name: &str,
-    expected_pkg: Option<&str>,
-    root: Option<&Path>,
-) -> Vec<Location> {
-    let mut cmd = std::process::Command::new("fd");
-    cmd.args([
-        "--type",
-        "f",
-        "--absolute-path",
-        "--max-results",
-        "10",
-        file_name,
-    ]);
-    if let Some(r) = root {
-        cmd.arg(r);
-    }
-
-    let out = match cmd.output() {
-        Ok(o) if o.status.success() => o,
-        _ => return vec![],
-    };
-    parse_fd_hits(&out.stdout, symbol_name, expected_pkg)
-}
-
-/// Parse a list of newline-separated absolute file paths from fd output,
-/// parse each file with the appropriate parser, and return locations for
-/// `symbol_name`.  When `expected_pkg` is given the package-exact match is
-/// returned immediately; otherwise the first match wins.  A non-exact match
-/// is kept as a fallback and returned only if no exact match is found.
-fn parse_fd_hits(stdout: &[u8], symbol_name: &str, expected_pkg: Option<&str>) -> Vec<Location> {
-    let mut fallback: Option<tower_lsp::lsp_types::Location> = None;
-
-    for path_str in String::from_utf8_lossy(stdout).lines() {
-        let path_str = path_str.trim();
-        if path_str.is_empty() {
-            continue;
-        }
-
-        let path = std::path::Path::new(path_str);
-        let Ok(uri) = tower_lsp::lsp_types::Url::from_file_path(path) else {
-            continue;
-        };
-        let Ok(content) = std::fs::read_to_string(path) else {
-            continue;
-        };
-
-        let file_data = parse_by_extension(path_str, &content);
-        let Some(sym) = file_data.symbols.iter().find(|s| s.name == symbol_name) else {
-            continue;
-        };
-
-        let loc = tower_lsp::lsp_types::Location {
-            uri,
-            range: sym.selection_range,
-        };
-
-        if let Some(pkg) = expected_pkg {
-            if file_data.package.as_deref() == Some(pkg) {
-                return vec![loc];
-            }
-            if fallback.is_none() {
-                fallback = Some(loc);
-            }
-        } else {
-            return vec![loc];
-        }
-    }
-
-    fallback.map(|l| vec![l]).unwrap_or_default()
 }
 
 /// Step 3 — same-package visibility (no import needed in Kotlin).
